Updated UI to handle SUV debugging, added limited tests, enhanced database viewing and editing

+++ WCI-SUV.Core/Interface/Database/IConveyorEntityService.cs
@@ -9,12 +9,12 @@ namespace WCI_SUV.Core.Interface.Database
 {
     public interface IConveyorEntityService
     {
-        Task<Conveyor> GetConveyorByIdAsync(long accountNumber);
+        Task<Conveyor> GetConveyorByIdAsync(Int32 accountNumber);
         Task<IEnumerable<Conveyor>> GetAllConveyorAsync();
         Task<Conveyor> AddConveyorAsync(Conveyor coneyor);
         Task UpdateConveyorAsync(Conveyor conveyor);
-        Task DeleteConveyorAsync(long accountNumber);
-        Task<bool> AccountNumberExists(long accountNumber);
-        public List<long>? GetSlotColumn();
+        Task DeleteConveyorAsync(Int32 accountNumber);
+        Task<bool> AccountNumberExists(Int32 accountNumber);
+        public List<Int32>? GetSlotColumn();
     }
 }
